Move fixture to base class
